完成首次测试。还有问题：1.管家名字2.消息文本间距（已经完成发送消息）3.删除最后一条图片消息失败4.收到图片消息，不能从缩略图直接取图片

+++ ChatVc/Model/MOrder.swift
@@ -15,7 +15,7 @@ class MOrder: TopModel {
     /// 类型名称
     var typeName : String = ""
     var goodsName = ""
-    var price : Float = 0
+    var price : Double = 0
     var status = 0
     var phone = ""
     var created = ""
